Move service::processor to runtime::service::instance::processor
This is part of a series of commits to move the existing APIs for
services into a new modules to make room for the new threadless design.
