+++ libsplinter/src/runtime/service/instance/mod.rs
@@ -12,14 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod factory;
-mod message_context;
-mod network_registry;
-mod network_sender;
-mod service_instance;
+mod processor;
 
-pub use factory::ServiceFactory;
-pub use message_context::ServiceMessageContext;
-pub use network_registry::ServiceNetworkRegistry;
-pub use network_sender::ServiceNetworkSender;
-pub use service_instance::ServiceInstance;
+pub use processor::{
+    registry::StandardServiceNetworkRegistry, JoinHandles, ServiceProcessor,
+    ServiceProcessorBuilder, ServiceProcessorError, ServiceProcessorShutdownHandle,
+};
