+++ services/scabbard/libscabbard/src/service/factory/mod.rs
@@ -37,7 +37,10 @@ use splinter::error::InternalError;
 use splinter::error::{InvalidArgumentError, InvalidStateError};
 use splinter::orchestrator::{OrchestratableService, OrchestratableServiceFactory};
 use splinter::service::validation::ServiceArgValidator;
-use splinter::service::{instance::ServiceInstance, FactoryCreateError, ServiceFactory};
+use splinter::service::{
+    instance::{ServiceFactory, ServiceInstance},
+    FactoryCreateError,
+};
 #[cfg(all(feature = "lmdb", any(feature = "postgres", feature = "sqlite")))]
 use transact::database::Database;
 #[cfg(any(feature = "postgres", feature = "sqlite"))]
