+++ splinterd/src/node/running/admin.rs
@@ -20,7 +20,7 @@ use splinter::events::Reactor;
 use splinter::peer::PeerManagerConnector;
 use splinter::registry::RegistryWriter;
 use splinter::rest_api::actix_web_1::Resource as Actix1Resource;
-use splinter::service::ServiceProcessorShutdownHandle;
+use splinter::runtime::service::instance::ServiceProcessorShutdownHandle;
 use splinter::store::StoreFactory;
 use splinter::threading::lifecycle::ShutdownHandle;
